--- server/api/v1/[...].ts
@@ -1,2 +1,0 @@
-import v1Router from '../../mvc/router'
-export default defineEventHandler(v1Router)
