+++ server/api/v1/forms/[formUlid]/index.get.ts
@@ -1,0 +1,25 @@
+import {getFormByUlid} from "~~/server/api/v1/forms/utils/queries";
+
+export default defineEventHandler(async event => {
+    const formUuid = getRouterParam(event, "formUlid")
+    if (!formUuid) return useHttpEnd(event, {
+        statusCode: Status.badRequest,
+        body: "No form ID provided"
+    }, Status.badRequest)
+
+    const form = await getFormByUlid(formUuid).catch(err => err as Error)
+    if (form instanceof Error) return useHttpEnd(event, {
+        statusCode: Status.internalServerError,
+        body: form?.message || "Unknown error while getting form"
+    } as APIResponse<string>, Status.internalServerError)
+    if (!form) return useHttpEnd(event, {
+        statusCode: Status.notFound,
+        body: "Form not found"
+    }, Status.notFound)
+
+    const response = {} as APIResponse<{ forms: Drizzle.Form.select, stores: Drizzle.Store.select }>
+    response.statusCode = Status.success
+    response.body = form as any
+
+    return response
+})
