+++ server/api/v1/forms/[formUlid]/submissions/index.get.ts
@@ -1,0 +1,27 @@
+import {getFormResponses} from "./utils/queries";
+
+export default defineEventHandler(async event => {
+    const formUlid = event.context.params?.formUlid
+    if (!formUlid) return useHttpEnd(event, {
+        statusCode: Status.badRequest,
+        body: "No form ID provided"
+    }, Status.badRequest)
+
+    const [details, error] = await useAuth(event)
+    if (error || !details) return useHttpEnd(event, {
+        statusCode: Status.unauthorized,
+        body: "Unauthorized"
+    })
+
+    const submissions = await getFormResponses(formUlid).catch(err => err as Error)
+    if (submissions instanceof Error) return useHttpEnd(event, {
+        statusCode: Status.internalServerError,
+        body: submissions.message || "Unknown error while getting form submissions"
+    }, Status.internalServerError)
+
+    const response = {} as APIResponse<typeof submissions>
+    response.statusCode = Status.success
+    response.body = submissions
+
+    return response
+})
