+++ server/api/v1/forms/mpesa/stk.ts
@@ -1,10 +1,8 @@
-import { insertFormPayment, insertPayment } from "../../mvc/forms/queries";
-import { H3Event } from "h3"
-import { insertWithdrawal } from "./queries";
+import {insertFormPayment, insertPayment} from "~~/server/api/v1/forms/utils/queries";
+import {H3Event} from "h3";
 
-const router = createRouter()
 
-function _getRequestIP(event: H3Event): string {
+export function _getRequestIP(event: H3Event): string {
     let ip = getRequestIP(event)?.toString()
 
     if (!ip) {
@@ -14,7 +12,7 @@ function _getRequestIP(event: H3Event): string {
     return ip || ""
 }
 
-router.use('/forms/callback/stk', defineEventHandler(async event => {
+export default defineEventHandler(async event => {
     const hook = await readBody(event) as StkCallbackHook
     const callback = hook.Body.stkCallback
     if (!callback) {
@@ -90,42 +88,4 @@ router.use('/forms/callback/stk', defineEventHandler(async event => {
     global.channels!.getChannel(channelName)?.terminate()
     log.success("Payment processed successfully Ref: " + callback.CallbackMetadata.Item.find(item => item.Name === "MpesaReceiptNumber")?.Value)
     return useHttpEnd(event, { statusCode: Status.success, body: "OK" })
-}))
-
-
-router.use('/forms/callback/withdrawal', defineEventHandler(async event => {
-    const body = await readBody(event) as B2cCallback
-    if (!body) {
-        return useHttpEnd(event, { statusCode: 400, body: "No callback found" }, 400)
-    }
-
-    const { Result } = body
-    if (!Result) {
-        return useHttpEnd(event, { statusCode: 400, body: "No callback found" }, 400)
-    }
-
-    const { ResultParameters, ResultType, OriginatorConversationID, ConversationID, TransactionID, ReferenceData } = Result
-    if (ResultType !== 0) {
-        log.error(`Failed to process withdrawal: ${Result.ResultDesc}`)
-        return useHttpEnd(event, { statusCode: 500, body: "Failed to process withdrawal" }, 200)
-    }
-
-    const transactionCode = ResultParameters.ResultParameter.find(param => param.Key === "TransactionReceipt")?.Value
-    const amount = ResultParameters.ResultParameter.find(param => param.Key === "Amount")?.Value
-
-    if (!transactionCode || !amount) {
-        log.error(`Failed to process withdrawal: ${Result.ResultDesc}`)
-        return useHttpEnd(event, { statusCode: 500, body: "Failed to process withdrawal" }, 200)
-    }
-
-    const ulid = await insertWithdrawal(+amount, transactionCode.toString()).catch(e => e as Error)
-
-
-    return useHttpEnd(event, { statusCode: 201, body: "OK" })
-}))
-
-router.use("/forms/timeout/withdrawal", defineEventHandler(async event => {
-    log.error("Withdrawal timed out", await readBody(event))
-}))
-
-export default useController('mpesa', router)
+})
