+++ server/api/v1/forms/mpesa/timeout.ts
@@ -1,0 +1,3 @@
+export default defineEventHandler(async event => {
+    log.error("Withdrawal timed out", await readBody(event))
+})
