+++ server/api/v1/forms/mpesa/utils/queries.ts
@@ -1,5 +1,5 @@
 import { withDrawals } from "~~/server/db/drizzle/schema";
-import db from "../../db";
+import db from "../../../../../db";
 import type { Drizzle } from "~~/server/db/types";
 
 export async function insertWithdrawal(amount: number, transactionCode: string) {
