+++ server/middleware/1.visit.logger.ts
@@ -1,4 +1,15 @@
-import {H3Event} from "h3";
-export default defineEventHandler((context: H3Event) => {
-    if(!isVercel) log.info(`[${context.node.req.method}]\t${context.node.req.url}`)
-})
+import type { H3Event } from "h3"
+
+function announcer() {
+    let time_start: number;
+    return {
+        handler(context: H3Event) {
+            time_start = performance.now()
+        },
+        onBeforeResponse(context: H3Event) {
+            if (!isVercel) log.info(`[${context.node.req.method}]\t${context.node.req.url} - ${(performance.now() - time_start).toLocaleString()}ms`)
+        }
+    }
+}
+
+export default defineEventHandler(announcer())
