+++ server/mvc/users/queries.ts
@@ -3,7 +3,7 @@ import db from "../../db";
 import {sessions, users} from "~~/server/db/drizzle/schema";
 import {and, eq} from "drizzle-orm";
 import {ulid} from "ulid";
-import { getRecentForms as _getRecentForms } from "../forms/queries";
+import { getRecentForms as _getRecentForms } from "../../api/v1/forms/utils/queries";
 
 export async function getUserByToken(token: string): Promise<Drizzle.User.select | null> {
     if (!token) return null
